[Test] fix Chatroom test cases to use factory method
